+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Test cases for all of the iterable mutation operators
   * Test cases for all of the mutation operators for permutations
   * Test cases for UniformMutation for integer representations
+  * Test cases for all of the classic annealing schedules (linear, exponential, and logarithmic)
 
 ### Dependencies
 * Bump core from 2.4.3 to 2.4.4